requirements.txt
arxiv==1.4.8
anthropic==0.43.1
pypdf==4.0.1
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
markdown==3.5.2
python-dotenv==1.0.1
requests==2.31.0 